requirements.txt
numpy==1.21.5
pmt==0.0.4
pycryptodome==3.23.0
PyQt5==5.15.11
PyQt5_sip==12.9.1
sip==6.14.0
